pyproject.toml
[tool.poetry]
name = "video-hosting"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "video_hosting"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
motor = "^3.4.0"
jinja2 = "^3.1.3"
uvicorn = "^0.29.0"
python-telegram-bot = "^21.0.1"
python-dotenv = "^1.0.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
